pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usability-teleop"
version = "0.1.0"
description = "Clean, reproducible pipeline for teleoperation usability research"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [
  { name = "Eduard" }
]
license = { text = "Proprietary (until release)" }
dependencies = [
  "numpy==1.26.4",
  "pandas==2.2.3",
  "scikit-learn==1.4.2",
  "scipy==1.17.1",
  "matplotlib==3.8.4",
  "seaborn==0.13.2",
  "xgboost==2.0.3",
  "shap==0.44.1",
  "PyYAML==6.0.3",
  "openpyxl==3.1.5",
  "tqdm==4.67.3"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-cov==6.0.0",
  "ruff==0.11.0",
  "mypy==1.15.0",
  "pandas-stubs==2.2.3.250308",
  "types-PyYAML==6.0.12.20241230"
]

[project.scripts]
usability-teleop = "usability_teleop.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
filterwarnings = [
  "ignore:'parseString' deprecated - use 'parse_string':DeprecationWarning",
  "ignore:'resetCache' deprecated - use 'reset_cache':DeprecationWarning",
  "ignore:'enablePackrat' deprecated - use 'enable_packrat':DeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
plugins = []
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = [
  "sklearn.*",
  "scipy.*",
  "xgboost",
  "xgboost.*",
  "seaborn",
  "seaborn.*",
  "shap",
  "shap.*",
]
ignore_missing_imports = true
